+++ src/index.ts
@@ -1,3 +1,3 @@
+export * from './basicAuthAuthenticator.js'
 export * from './publish.js'
 export * from './types.js'
-export * from './vercelAuthenticator.js'